requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
httpx>=0.27.2
python-dotenv>=1.0.1
anthropic>=0.39.0
pydantic>=2.11.0
beautifulsoup4>=4.12.3
